Update oauth2login README with GitHub/Facebook/Okta
* Add instructions to the oauth2login sample README for setting up
OAuth 2.0 Login for GitHub, Facebook, and Okta
* Change base package for oauth2login sample

Fixes gh-4287, Fixes gh-4288, Fixes gh-4289
